# Правки багов

+++ src/main/java/com/example/tgbot/models/configurations/ModelRequestOptions.java
@@ -5,7 +5,7 @@ import com.example.tgbot.models.enums.GenerationModel;
 import java.util.Map;
 
 public interface ModelRequestOptions {
-    double getPrice();
+    int getPrice();
     Map<String, Object> getRequestInput();
     String getOptionsText();
     void setParametersFromJson(String json);
